Refactor heatmap selector and configuration
- Removed unused layout types from HeatmapSelectorTemplate.
- Updated WebHeatmapDefaults to streamline configurations and removed redundant methods.
- Enhanced UniversalHeatmapWidget with improved logging and performance metrics.
- Introduced SectorHeatmapConverter to convert portfolio analytics data into generic heatmap data.
- Simplified heatmap data handling and improved readability across multiple files.

+++ lib/features/portfolio/presentation/mappers/sector_heatmap_converter.dart
@@ -1,6 +1,6 @@
-import '../../features/portfolio/internal/domain/entities/portfolio_analytics.dart';
-import '../models/heatmap.dart';
-import '../widgets/heatmap/heatmap_config.dart' as ui_config;
+import '../../../../shared/models/heatmap.dart';
+import '../../../../shared/widgets/heatmap/heatmap_config.dart' as ui_config;
+import '../../internal/domain/entities/portfolio_analytics.dart';
 
 /// Utility class to convert portfolio analytics data to generic heatmap data
 class SectorHeatmapConverter {
@@ -58,7 +58,7 @@ class SectorHeatmapConverter {
             id: sector.sectorName,
             name: sector.sectorName,
             displayName: _getSectorDisplayName(sector.sectorName),
-            weightage: weightage,
+            weightage: weightage.toDouble(),
             performance: avgPerformance,
             value: sectorValue,
             metadata: {
